added missing translations, moved initizaliton of Babel from Flask entry point into __init__ of Flask App for better code structure
