+++ app/static/translations/de/LC_MESSAGES/messages.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2021-03-12 16:22+0100\n"
+"POT-Creation-Date: 2021-03-12 17:04+0100\n"
 "PO-Revision-Date: 2021-03-12 13:53+0100\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language: de\n"
@@ -49,92 +49,6 @@ msgstr "Gespeicherte Skizzen"
 #: app/templates/index.html:119
 msgid "Completed Experiments"
 msgstr "Abgeschlossene Experimente"
-
-#: app/templates/layout.html:33
-msgid "Home"
-msgstr "Startseite"
-
-#: app/templates/layout.html:39
-msgid "Interface"
-msgstr "Interface"
-
-#: app/templates/layout.html:46
-msgid "Create..."
-msgstr "Erstelle..."
-
-#: app/templates/layout.html:51
-msgid "Create a new:"
-msgstr "Erstelle ein neues:"
-
-#: app/templates/layout.html:52
-msgid "Project"
-msgstr "Projekt"
-
-#: app/templates/layout.html:53
-msgid "Experiment"
-msgstr "Experiment"
-
-#: app/templates/layout.html:54
-msgid "Sketch"
-msgstr "Skizze"
-
-#: app/templates/layout.html:62
-msgid "Projects"
-msgstr "Projekte"
-
-#: app/templates/layout.html:68
-msgid "Experiments"
-msgstr "Experimente"
-
-#: app/templates/layout.html:74
-msgid "Pages"
-msgstr "Account"
-
-#: app/templates/auth/profile.html:2 app/templates/layout.html:80
-msgid "Profile"
-msgstr "Profil"
-
-#: app/templates/auth/login.html:1 app/templates/auth/login.html:8
-#: app/templates/auth/login.html:32 app/templates/layout.html:88
-msgid "Login"
-msgstr "Anmelden"
-
-#: app/templates/auth/register.html:1 app/templates/auth/register.html:44
-#: app/templates/layout.html:93
-msgid "Register"
-msgstr "Registrieren"
-
-#: app/templates/layout.html:99 app/templates/layout.html:171
-msgid "Logout"
-msgstr "Logou"
-
-#: app/templates/layout.html:107
-msgid "Current Language"
-msgstr "Aktuelle Sprache"
-
-#: app/templates/layout.html:114 app/templates/layout.html:127
-msgid "English"
-msgstr "Englisch"
-
-#: app/templates/layout.html:119 app/templates/layout.html:130
-msgid "German"
-msgstr "Deutsch"
-
-#: app/templates/layout.html:125
-msgid "Change language to:"
-msgstr "Sprache ändern zu:"
-
-#: app/templates/layout.html:163
-msgid "Ready to Leave?"
-msgstr "Du willst dich ausloggen?"
-
-#: app/templates/layout.html:168
-msgid "Select \"Logout\" below if you are ready to end your current session."
-msgstr "Klicke unten auf \"Logout\", um dich auszuloggen."
-
-#: app/templates/layout.html:170
-msgid "Cancel"
-msgstr "Abbrechen"
 
 #: app/templates/auth/change-building.html:1
 msgid "Change User Building"
@@ -193,6 +107,11 @@ msgstr "Ändere deinen Standort"
 msgid "Your current Site is:"
 msgstr "Dein aktueller Standort ist:"
 
+#: app/templates/auth/login.html:1 app/templates/auth/login.html:8
+#: app/templates/auth/login.html:32
+msgid "Login"
+msgstr "Anmelden"
+
 #: app/templates/auth/login.html:35
 msgid "Need an account?"
 msgstr "Du brauchst einen Account?"
@@ -200,6 +119,10 @@ msgstr "Du brauchst einen Account?"
 #: app/templates/auth/login.html:35
 msgid "Signup!"
 msgstr "Registrieren!"
+
+#: app/templates/auth/profile.html:2
+msgid "Profile"
+msgstr "Profil"
 
 #: app/templates/auth/profile.html:14
 msgid "Your Profile"
@@ -244,6 +167,10 @@ msgstr "Aktuelles Gebäude ändern"
 #: app/templates/auth/profile.html:89
 msgid "Change Current Room"
 msgstr "Aktuellen Raum ändern"
+
+#: app/templates/auth/register.html:1 app/templates/auth/register.html:44
+msgid "Register"
+msgstr "Registrieren"
 
 #: app/templates/auth/register.html:8
 msgid "Sign up"
@@ -302,7 +229,7 @@ msgstr "Erstelle ein Experiment"
 
 #: app/templates/forms/create-project.html:37
 msgid "Create Project"
-msgstr "Erstelle ein Projelt"
+msgstr "Erstelle ein Projekt"
 
 #: app/templates/interface/choose-project.html:8
 msgid "Choose a Project to create an Experiment for"
@@ -546,4 +473,70 @@ msgid "You can not update this experiment - This ELN Number is already in use!"
 msgstr ""
 "Du kannst dieses Experiment nicht aktualisieren - Diese ELN-Nummer ist "
 "bereits in Verwendung!"
+
+#~ msgid "Home"
+#~ msgstr "Startseite"
+
+#~ msgid "Interface"
+#~ msgstr "Interface"
+
+#~ msgid "Create..."
+#~ msgstr "Erstelle..."
+
+#~ msgid "Create a new:"
+#~ msgstr "Erstelle ein neues:"
+
+#~ msgid "Project"
+#~ msgstr "Projekt"
+
+#~ msgid "Experiment"
+#~ msgstr "Experiment"
+
+#~ msgid "Sketch"
+#~ msgstr "Skizze"
+
+#~ msgid "Projects"
+#~ msgstr "Projekte"
+
+#~ msgid "Experiments"
+#~ msgstr "Experimente"
+
+#~ msgid "Pages"
+#~ msgstr "Account"
+
+#~ msgid "Logout"
+#~ msgstr "Logou"
+
+#~ msgid "Current Language"
+#~ msgstr "Aktuelle Sprache"
+
+#~ msgid "English"
+#~ msgstr "Englisch"
+
+#~ msgid "German"
+#~ msgstr "Deutsch"
+
+#~ msgid "Change language to:"
+#~ msgstr "Sprache ändern zu:"
+
+#~ msgid "Ready to Leave?"
+#~ msgstr "Du willst dich ausloggen?"
+
+#~ msgid "Select \"Logout\" below if you are ready to end your current session."
+#~ msgstr "Klicke unten auf \"Logout\", um dich auszuloggen."
+
+#~ msgid "Cancel"
+#~ msgstr "Abbrechen"
+
+#~ msgid "Update this Project!"
+#~ msgstr "Dieses Projekt updaten!"
+
+#~ msgid "Update Project"
+#~ msgstr "Projekt updaten"
+
+#~ msgid "Update this Experiment!"
+#~ msgstr "Dieses Experiment updaten!"
+
+#~ msgid "Update Experiment"
+#~ msgstr "Experiment aktualisieren"
