+++ app/util/extensions.py
@@ -3,6 +3,7 @@ from flask_jsglue import JSGlue
 from celery import Celery
 from flask_bootstrap import Bootstrap
 from flask_babel import Babel
+
 babel = Babel()
 db = SQLAlchemy()
 jsglue = JSGlue()
